% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_int.R
\name{dat_int}
\alias{dat_int}
\title{interpolate data}
\usage{
dat_int(band_data, band_name)
}
\arguments{
\item{band_data}{data to be interpolated}

\item{band_name}{band name}
}
\value{
interpolated daily time-series data
}
\description{
interpolate data
}
\examples{
library(GMothDef)
library(stringr)
setwd('C:/YourFolder')
VV <- read.csv(file = 'VV.csv')
dat_int(VV, 'VV')
}
