% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_smth.R
\name{dat_smth}
\alias{dat_smth}
\title{Loess Smoothing}
\usage{
dat_smth(dat, span)
}
\arguments{
\item{dat}{daily data to be smoothed e.g. output from dat_int()}

\item{span}{parameter of smoothing}
}
\value{
smoothed data
}
\description{
Loess Smoothing
}
\examples{
library(GMothDef)
dat_smth (s1_VV_t, 0.75)
}
