% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_tiff.R
\name{dat_tiff}
\alias{dat_tiff}
\title{Data extraction from Tiff}
\usage{
dat_tiff(band_names, plot_shp, shp_type, date_start, ID_plot)
}
\arguments{
\item{band_names}{a vector of band names}

\item{plot_shp}{shapefile of the plots}

\item{shp_type}{'point' or 'polygon}

\item{date_start}{date start in the file name}

\item{ID_plot}{the field name of plot id}
}
\value{
files saved in the working directory
}
\description{
Data extraction from Tiff
}
\examples{
library(GMothDef)
library(gtools)
library(raster)
library(rgdal)
library(TDPanalysis)
setwd("C:/YourFolder")
band <- c('VH', 'VV')
dat_tiff(band, 'study_area.shp', 'point', 25, 'ID')
}
