% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_dat.R
\name{predict_dat}
\alias{predict_dat}
\title{Make predictions using rf model}
\usage{
predict_dat(rf_model, newd, ID_plot, ID_pred)
}
\arguments{
\item{rf_model}{the model to make predictions}

\item{newd}{new data for predictions}

\item{ID_plot}{the field name of plot names in the new data (newd)}

\item{ID_pred}{the field/file name to record prediction data}
}
\value{
prediction
}
\description{
Make predictions using rf model
}
\examples{
library(GMothDef)
predict_dat(rf, newd, 'id', 'SR_pred')
}
