% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_reg.R
\name{rf_reg}
\alias{rf_reg}
\title{Random Forest Regression}
\usage{
rf_reg(x, y)
}
\arguments{
\item{x}{predictor variables}

\item{y}{response variable}
}
\value{
the best model
}
\description{
Random Forest Regression
}
\examples{
library(GMothDef)
rf_reg(v_pred, v_res)
}
