% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smth_span.R
\name{smth_span}
\alias{smth_span}
\title{choose smoothing span}
\usage{
smth_span(band_name, dat, span)
}
\arguments{
\item{band_name}{band name/title of the output plot}

\item{dat}{daily data to be smoothed e.g. output from dat_int()}

\item{span}{a vector of spans to be compared}
}
\value{
a plot to help choosing smoothing span
}
\description{
choose smoothing span
}
\examples{
library(GMothDef)
library(stringr)
library(dplyr)
library(tidyr)
library(ggplot2)
setwd('C:/YourFolder')
VV <- read.csv(file = 'VV.csv')
dat_int(VV, 'VV')
smth_span('VV', dat_int, c(0.25,0.5,0.75))
}
